refactor(fuzz): split oversized mod.rs and types.rs (#6166)
